made whereToSnap vertically and horizontally snap on closest
before it snapped and returned the first one it snapped.
Now tries all, gets the distance to the center of snapped area on each and gets the smallest one.
This way it no longer depends on the order of snap checks where it snaps.
Now it always snaps to the smallest one only